[watchman] add integration test machinery
Summary:
in order to validate the server functionality, we need
to be able to define integration tests that run the server and
make it do things.  This diff defines an integration test engine
that uses the arcanist test infrastructure for this purpose,
and adds an example integration test case.

These are not included as part of "arc unit" because they take
longer to run (especially since we have a minimum 1second delay
in the client at the moment due to the poor EOF detection code
there).

Test Plan: `make integration` or `arc unit --engine WatchmanIntegrationEngine`

Reviewers: jakubv, nharper, keegancsmith, zoel

Differential Revision: https://phabricator.fb.com/D645447