[watchman] jsonify protocol
Summary:
fix the protocol handling by adopting json for both the requests and
the response.

This also allows us to remove the dependency on libevent (which is honestly a
bit of a cluster... to handle portably), so now all our dependencies are either
in the base system or are included in our source code.

This depends on D645447 for the integration test pieces.

We always return a JSON object response that includes our version number.
Errors will always have an "error" object field that holds the text of the
error.  The rest of the commands are still transitional when it comes to their
response format, but I'll be updating the README as I review/revise each of
those.

I've pulled in version 2.4 of the Jansson JSON library, which is compatibly MIT
licensed.  Doc on it can be found here: http://www.digip.org/jansson/doc/2.4/

Test Plan: `make integration`

Reviewers: jakubv, keegancsmith, zoel

CC: pieter, fugalh, security-diffs@lists

Differential Revision: https://phabricator.fb.com/D645534