[watchman] adopt json arguments for watchman commands
Summary:
build on D645534 and use the structured data we parse
from the json requests directly in the argument handling of
the watchman command dispatcher.

This allows us to remove some code for argv manipulation
but also paves the way for using a more structured representation
for some of the command parameters.

Test Plan: `make integration`

Reviewers: jakubv, keegancsmith, zoel, wez

Reviewed By: wez

CC: pieter, scottg

Differential Revision: https://phabricator.fb.com/D645596