[watchman] add integration test for the trigger command
Summary:
add a test for the "trigger" command, which is used
to spawn a process and invoke it when files that match a pattern
are observed to change.

This builds on the debug connection from the prior diff to
get its stuff done, and introduces a couple of nice helpers
functions that reduce flaky results without too many arbitrary
pauses.

Test Plan:
`arc unit`.
Note that there is a race in basic.php that can trigger if
we get multiple notifies about the file that we touch; these
can cause the cursor to return an additional file at the point
where we expect it to usually return none.

Reviewers: keegancsmith, dsereni, zoel, pieter

Reviewed By: zoel

Differential Revision: https://phabricator.fb.com/D648760