Add concept of a settle delay
The settle delay is used in two places:

 * We'll delay triggers by the settle delay time, measured from
   the last observed file change time.
 * We'll force clients that are reading changes via a named
   cursor to wait for the settle delay time to pass before
   returning their results.

If file changes are detected while waiting for the settle delay
to pass, we'll reset and wait for the settle delay time after this
latest change.

This may mean that we never trigger or satisfy the change condition
if the location is always being updated... we'll have to see if that
poses a problem in practice.

These changes appear to have smoothed out the race in the testCursor
unit test.

You may set the settle time via the --settle option.  It is specified
in milliseconds.  One does not simply sleep for less than 1 ms:
http://wezfurlong.org/blog/2011/may/short-sleeps-on-solaris/