add test for renames, and deletes + recreates
Found that I wasn't consuming inotify events properly, leading to
flaky test results.

Belt-and-suspenders this a bit more in the test harness.  This
makes the trigger test case apparently more flaky, as we observe
more inotify events now.  Might have to make the settle time
variable on a per watch basis to event this out.