Limit trigger process concurrency to 1
Move the child-waiting into a dedicated thread so that:

 * We are not limited to async-signal-safe functions in the handler
 * Do not have to attempt to lock a root from the wrong context if
   we get EINTR'd

This makes it much simpler; when a process is done, we want to
re-evaluate whether any files have changed since it ran, if so,
then we check the set of matches and see if we should spawn it
again.

This change means that we don't fork bomb ourselves in the face
of a large number of file updates, while ensuring that we do
actually spawn the process and don't miss updates.