triggers: limit arg size and provide json input
Respect the system ARG_MAX limit for command line length when we
generate an invocation to a trigger command.

Also provide the file list as a JSON array on its stdin.

This means that consuming stdin will be the most reliable and complete
information available, as we will stop adding args to the argv if we
would go over the limit on argument length.

We can't *simply* split our command invocations across two or more
spawns because we can't retain our matched file list in any sane
way across those invocations; files may be changing at any moment,
or even be changed by the process we're spawning.