include dirs in the changed file list
I realized that we could track and report on dirs if we simply
treat all filesystem entries as nodes.  We still maintain a tree
of watchman_dir structs, but now we also include IS_DIR entries
in the "file" list and they show up like everything else.

Also flesh out some test cases and bump in preparation for 1.2.