ignore VCS control dirs
Our huge repos can cause a large amount of extra work for ourselves
by keeping us busy looking at their control data.  This hard-codes
(for now) ignores for those dirs so that we can start up and process
matches faster.

Will be making the ignores configurable via the `watch` command in
a follow-on commit, but want to get this deployed to our dev tier
first.