First pass at new query engine
This is far enough along that it can be played with.
There are no tests for it yet, and it currently lacks
expressions for comparing modification times (most useful
if you're not using the 'since' generator) or comparing
integer properties of the file.

I think this is good enough to stick in 1.5 under a beta
label.  This needs to get integrated into triggers and
we'll use this as the basis of subscription support also,
and then re-implement 'since' and 'find' in terms of this
new engine.