Add "clientmode" for some commands
In some cases we don't want to start the service but don't want to
code alternate logic using `find` or other shell or system functions.

This commit adds a `--no-spawn` option that changes the CLI behavior
so that it will try to communicate with a running watchman service,
but if it is not running will fall back to a set of client mode
commands and execute those.

At this point in time we only have the new `query` command available;
it works by crawling the requested root and then executing the query
over it.  It will only have sane results for the `path` generator.