Add subscribe & unsubscribe facilities
Summary:
This adds subscription support to watchman.

Subscriptions are scoped to a given client connection, and further
scoped to a watched root and a client-specified subscription name.

Subscriptions may be explicitly cancelled via the unsubscribe command
and are implicitly cancelled when the client disconnects.

Subscriptions are built on top of the new query API that is included
in watchman 1.5 that provides more filtering of the returned data,
as well as specifying the returned fields:

https://github.com/facebook/watchman#command-query-beta-starting-in-v15
https://github.com/facebook/watchman#file-queries

(but note that this diff fixes a refcnt issue; don't use the query command
in volume with 1.5!)

When a subscription is activated on a connection, watchman will unilaterally
send `subscription` data packets to the client as matching changes are
detected.  The initial subscription request can be scoped to a given `since`
value (or leave it unspecified) and will send a subscription packet immediately
following the subscription request if any files were matched.

This frees clients from having to schedule a recurring `since` poll against the
service.

Test Plan:
`arc unit` and `WATCHMAN_VALGRIND=1 arc unit` to prove functionality
and that there are no leaks.

The test case is rather anemic... move fast etc. etc.

Reviewers: mroch, pieter, lifeihuang, wez

Reviewed By: wez

Differential Revision: https://phabricator.fb.com/D707379