Be more proactive at detecting and healing inconsistencies
Summary: Add a should_recrawl flag that we set when we detect
inconsistencies.  At the moment this is limited to error
conditions around inotify usage; when we detect those we set
the flag.  At defined "safe" points we check this flag and
queue up a full tree recrawl to make sure we're in a good
state.

These safe points are at the bottom of the inotify loop
and inside the stat thread loop.

Test Plan: `arc unit` shows that we didn't totally break it,
but the thing we're trying to fix is a very hard to repro
race condition.