Fixup opendir() path string
Well, this was embarrasing.  Turned out that we could pass a sliced
string to opendir.  This meant that if we got notified about a path
"foo/bar" and then decided to look at dirname("foo/bar") we'd think
we were calling opendir("foo") but were actually invoking
opendir("foo/bar").

This meant that we could prune out all the entries of "foo" that we
didn't see in "foo/bar", which meant that they'd show up deleted.
Whoops!

Also be a little more conservative when we get notified about deletes
and moves of watched firs; schedule a crawl of their parents.

And finally: explicitly stopping watches can result in race conditions,
so we simply NOP in stop_watching_dir(); Linux will notify us when
the watch descriptor is invalidated.