Add support for stopping watches
It seems that a common conflict resolution strategy is to simply
delete a watched tree and clone it again.  This can cause things
to go awry, so we needed to handle it better.

This diff makes it possible to list and delete watches, and to
automatically delete watches when the watched directory is removed.

To support this change, roots are now reference counted.
I also had to adjust the test harness code to better collect data
from valgrind and prove that we don't leak.