Introduce periodic polling
We keep running into an issue where we don't always see both sides
of move or fast create/delete operations for a given file, resulting
in it showing up as exists:false.

This is the first step at avoiding this class of issue; once the
filesystem has settled, instead of sleeping until the next notifcation,
we'll set an upper bound on the sleep and then when we wake up we'll
speculatively recrawl the tree.

I'm going to follow up with a couple of diffs that make this delay
more configurable, and potentially combine the crawl with the first
notification post-settle, although I want to be careful about scheduling
multiple full-tree crawls throughout a busy period.