Add a system-wide configuration file
This is a json file that is read before command line arguments
are parsed.  This has two important implications:

- We can't pass the path to the config file via the arg parser
- command line args override options from the config file

Change how we pass options to the service when running via gimli;
rather than materializing options and forcing them into the argv
array, we only pass them on if they were present in the originating
argv array.  This allows us to default parameters to the config
file without killing the gimli monitor processes.