Fix recursive crawl case
Another installment in well-that's-embarrassing: the recursive
directly crawler was only running in cases where a given item
was newly created or deleted.  That meant that
2399d745c7c0c2c1c925fb211c9e0a46198fd269 was only partially
effective.

In running this down I've decided to default the period crawl
to 0, and to make a 0 value turn it off; I'd seen that this
was causing some high CPU utilization quite frequently due
to the size of our tree.