fix nasty wd -> name state issue
Summary:
In the case of moves, it was possible for our mapping of watch descriptor => path name to become out of date.

This diff addresses this issue by invalidating the source of a move before we pass it in for processing in the stat thread.

Test Plan:
replace the local watchman binary with the older version; run
`arc unit` to repro the failure.
Then make with this version and the test passes.

Reviewers: sid0

Reviewed By: sid0

CC: arice

Differential Revision: https://phabricator.fb.com/D790536

Task ID: 2336595, 2334495