Handle sysctl limit constraints better
We had a situation in a test suite where we exceeded
`/proc/sys/fs/inotify/max_user_instances`.  Watchman would blindly
assume that it had a working watch but where it was not actually
plumbed.

This makes the error condition get logged to the log, and the watch
command will return "unable to resolve root", which is generic but
good enough for now.