Don't issue spurious recrawl for moves
Summary: 4f89e886828ef0d introduced this issue; when a dir is moved
we receive IN_MOVE_SELF but because we deleted our mapping we fall
into the something fishy code path.

Since there isn't any actionable information for us in IN_MOVE_SELF
(we have no way to know what path it is referring to),
we can ignore it and NOT consider it fishy.  We either have or will
receive a notification about its destination.