Revise VCS ignore behavior
Summary:
It turns out that we don't really want to completely ignore
the VCS control dirs because it makes it difficult to detect the
boundaries between nested repos and difficult to detect when a rebase
operation is in progress.  This diff changes our ignore behavior
so that we'll track the VCS control dir and its immediate contents,
but we don't recurse into child dirs.

When it comes to triggers, we may need to add some special knowledge
of the .hg/wlock or the .git/index.lock file mechanics to avoid starting
the trigger.  I don't really want to do that; need to do some
experiments to see if we need to go that far.

Test Plan: `arc unit`

Reviewers: sid0

Reviewed By: sid0

CC: vkatich

Differential Revision: https://phabricator.fb.com/D791827

Task ID: 2334371