use cookie files for view synchronization
Summary:
So that we can be sure to have observed changes that took place
at or before "now" we generate a cookie file somewhere in the tree.
Once we've observed it via inotify, we know that we're current.

Pick dir under .git, .hg or .svn as a first choice so that we avoid
showing up in the working copy state.

Test Plan:
`arc unit --everything` and `WATCHMAN_VALGRIND=1 arc unit --everything`

Reviewers: sid0

Reviewed By: sid0

Differential Revision: https://phabricator.fb.com/D800567

Task ID: 2348274