fixup cursor test case for OSX
Use our less racy assertFileListUsingSince to validate its behavior.
I have mixed feelings about this because of its timing sensitivity.

Also: we were not noticing that 'one' had changed because we either
didn't decide that we needed to lstat it again, or because the touch
call didn't have a visible effect (all happens in the same second).