remove stat thread from linux port
Summary:
This replaces the stat thread with "inline" FS operations on the
inotify thread.  This is a result of realizing that the mutex we use to
lock the root effectively serializes this effort anyway.  (An earlier
design was looking at using rwlocks, but it was realized that the volume
of clients would typically be very small).

This diff refactors the inotify processing so that it can be carried out
in some clearer steps:

 * wait for up to settle period for inotify events
 * if none are received, process triggers
 * otherwise, consume the inotify data from the kernel
 * enqueue the work
 * (we might consider looping here if we want to be aggressive)
 * dispatch the queued work; stat and update the tree

I like how the resulting code looks, and feel happy that this avoids
a potential race condition from the old model with the processing being
done more incrementally, punctuated by releasing and re-acquiring locks.

Bonus: can eliminate some condition variable signaling

There are a couple of things that will manifest differently:

 * we don't literally check the settle timeout before running triggers, so we may spuriously trigger too soon
 * this disables the unused recrawl code
 * the initial crawl of the tree will now block all queries until it completes (previously we'd unlock periodically).  Since this is a tree traversal, this won't be unbounded, but malicious code that creates an infinitely deep tree can make us hurt more than we used to.

I want to make equivalent changes for the other platforms, but this diff is big enough at the moment.

Test Plan: `arc unit --everything`

Reviewers: sid0

Reviewed By: sid0

Differential Revision: https://phabricator.fb.com/D805979