Refactor the removal of the stat thread and port it
This completely removes the stat thread and adjusts the BSD and Solaris
ports to use the same model as 0df6518027bd

In doing this, we can refactor the main notify thread loop so that there
is just one very clear set of steps for handling incoming events.

This diff also fixes the directory traversal flags on the solaris port
so that we don't needlessly crawl directories in response to cookies
being created/deleted in sync-to-now.