Raise the default sync_timeout to 1 minute
This is really my preferred default timeout.
I'm seeing some queries timeout in some early startup scenarios in our
build; I believe this is related to the amount of effort required to
dispatch triggers from the initial startup with a large number of files;
it takes us just over the edge of the low timeout that we had.