Improve handling of the root-went-away cases
Add explicit handling for these on BSD and Solaris and avoid
races in the test cases that are sensitive to it.

Related: if we decide to recrawl, drain any pending entries and
just crawl again from the root.  This is to avoid the risk that
we'll waste time looking for that stuff multiple times and then
trigger more queue overflow events.