Be more aggressive at consuming notifications
The goal is to avoid hitting an inotify queue overflow in
reasonably common scenarios.

This causes us to consume notifications while they are available
or until we have >= 16k to process.

We'll batch these up but uniq them in the pending list.

This centralizes the batching code (stealing it from the BSD port)
and avoids some alloc/free/realloc cycles.

Move the BSD and Solaris event buffers to be heap allocated and
stored in the tail end of the root struct just like we do for Linux.