Fix use-after-free with triggered processes
I think this bubbled up with the increase sleep period in the
reaper.  Most of this diff is adding instrumentation to help
track down the problem; W_LOG_FATAL is equivalent to W_LOG_ERR
but will print a stack trace if the system supports it (turns out
that Linux, OSX and Solaris all provide the backtrace() function)
and then abort.

The referencing problem was that the running child process should
have been addref'd and delref'd at the appropriate points, so
do that.

Also noticed that we should have locked the root around dispatching
triggers and subscriptions, so do that too.