BSD: remove file watching code; only watch dirs
I've been toying with this for a bit; I don't like that kqueue
requires one fd per watched item, but since it can notify on
changes in dirs, we don't have to watch the files.

So: don't do that.  Remove the file watching code and simplify.
Passes the test suite.