Bring back the reaper thread
I've seen waitpid(WNOHANG) block on the listener, leading to watchdog
kill (and no service while blocked).

I'm somewhat nonplussed that WNOHANG can hang, which leads me to suspect
some subtle unix detail that is not obvious from the various manpages.

Reinstate the reaper thread to handle this in a safer context.