add .watchmanconfig
Allow setting root specific options in a json file named .watchmanconfig
in the root of the root.

Use this to specify the settle period rather than via the command line.
This makes it much simpler to establish watches from multiple clients
without forcing them to invoke watchman with the same settle parameter.