Turn off non-blocking in BSER reader
The JSON reader handles this reasonably (we bubble back up to the poll)
but the logic in the bser handler is a little bit more brittle at
the moment.  Just turn on blocking to make it work correctly until
I can make a second pass over it.

This manifested as an intermittent failure on OSX, presumably due
to differences in how bufs are managed in the kernel making it more
likely that we'd EAGAIN; would pass reliably on Linux.