Fixup compilation and tests on Solaris
When we removed the 64-bit compilation hammer we started to fail
some of our tests because we're running them them with a 64-bit PHP
but a 32-bit watchman.

So, force it back on to 64-bit when building on Solaris (why wouldn't
you want that anyway?).

Also adjust the pcre test to allow for the regex to be parsed
successfully.