capture config, fail faster in test
Copy the config file into the log so we can see how the instance
was configured.

Ping the instance to get its pid before starting a test; if dead,
stop running tests.  This should cause us to fail fast and to highlight
failures easier: we have a lot of log data to crawl through.