refactor command registration and dispatch
This is a step towards some enhancements I'm planning for the watchman
CLI process.

Switch the registration from a central list of commands (and a separate
list for the client mode command set) to declaring the registration
details alongside the command function with some flags to indicate
whether it is permitted in client or daemon mode.

This can later be extended to include a CLI-side command preprocessor
callback that can do things like `realpath`ing the root dir on the
client side before sending a watch command.

We could also extend this to include a usage summary to be displayed
when `watchman -h` is invoked.

As part of this refactor, I took out the function call that initialized
the query expression term hash, so I applied the same principle to that
subsystem.