preprocess commands in the CLI
Adds a per-command callback that is invoked by the CLI prior to sending
a command to the server.

For commands that accept the root dir as the second parameter, we wire
up a callback that executes `realpath` on the root prior to sending the
command.  This makes it more convenient to use the CLI, as you may now
specify relative paths for the root.

This does not apply or have meaning to clients that are directly
accessing the socket, because they are bypassing the CLI.

References https://github.com/facebook/watchman/issues/33