add illegal_fstype global configuration
Summary:
Every so often we have to troubleshoot what looks like hung, stalled
or timed-out watchman processes.  These are typically caused by someone
watching an NFS mounted dir.   In the past, we tried to gate the software that
was setting up the watches to prevent it from launching on NFS, but that gating
isn't 100% foolproof.

This diff adds a pair of configuration options that can specify a set of
blacklisted filesystem type names and the site specific advice on what to
do if the dir they plan to watch is on the blacklist.

More details are in the README

Test Plan:
`make integration`.  Also manually attempted a watch in my VMs
(those mount my laptop via NFS) to confirm that they detect the filesystem type
names.

Reviewers: sid0

Reviewed By: sid0

CC: donguo, lifeihuang, joelb

Differential Revision: https://phabricator.fb.com/D1323424