Make another attempt to appease Travis
The Travis WebLint (http://lint.travis-ci.org/) complains if you pass
more than one language value, so it takes the first and ignores the
others.

We were relying on this to install PHP, but I think it means other
settings in `.travis.yml` are ignored; the linter reports:

  specified compiler, but setting is not relevant for php
  specified python, but setting is not relevant for php

So, make PHP the only language (so that it still gets installed) and try
installing Ruby and Bundler by other means. I'm leaving the other
settings in because there are apparently plans for Travis to eventually
support multiple languages within a single project:

https://groups.google.com/forum/#!topic/travis-ci/3pzCgvxWZAo