refactor python bser module to pywatchman.bser, add pywatchman.client
Summary:
This draws from our hgwatchman client code to provide a simple
watchman client API for python code.

This changes the makefile so that we install our watchman python module
during `make install`.  It also wires up a rudimentary test of the
client that talks to one of the instances set up by
WatchmanIntegrationEngine.

Test Plan: `./configure --with-python ; make integration`

Reviewers: sid0, achao

Reviewed By: sid0

Subscribers: moshez

Differential Revision: https://phabricator.fb.com/D1323720