When adding a trigger, bump the tick count
In the case where you add a trigger but don't change any files,
we were quietly sitting on the total list of files until the next
observed file change.

This makes trigger work a little more like subscribe (which sends you
your initial results when you subscribe), and makes it easier when
you're setting up triggers: you won't need to register the trigger and
then run the build command anyway.

Also: our docs say that it works like this, but it hasn't for a while.