age-out cursors when aging out file nodes
Summary:
Cursors that have a tick value lower than the last gc age
effectively have a 0 value, so we can safely remove them from
the cursor hash.

This supports a use case where tooling generates a unique cursor name on
startup and re-uses it for a while.

This works "best" when the rate at which the tool starts up is
relatively low (a handful of times a day at most).

Test Plan: `make integration`

Reviewers: sid0

Reviewed By: sid0

Subscribers: jimp

Differential Revision: https://phabricator.fb.com/D1413547