revise travis build so that we can run osx and linux
This also pulls in hgwatchman so that we can run its test suite.
I've disabled this for the moment because it takes a while on osx
and errors out :-/

We need to resolve that, but in the meantime, there's no reason not to
proceed with the build fixed up.

Also note: this turns off valgrind on travis builds too; we can revisit
and enable in a follow-on diff