Improve reporting on recrawls
Summary:
Adds a `warning` field to the response struct.  If present, it conveys
a warning that should be reported back to the user.

We currently only set this warning state if a recrawl is triggered.  The
warning output links to new documentation about what a recrawl is and how to
avoid it.

We maintain a recrawl counter but don't expose this in a structured way to the
client.  Not sure how big a deal this really is; given @sid0's recent change to
heal for non IN_Q_OVERFLOW conditions, this should only trigger when the
inotify resources are not tuned properly and thus this should be pretty much a
binary condition.

Test Plan:
`make integration`; tests that trigger recrawls assert that we set
the warning info.

The warning field looks like this:

```
Recrawled this watch 1 times, most recently because:
/private/var/tmp/bwx0ydi87bscgkc0: debug-recrawl
To resolve, please review the information on
https://facebook.github.io/watchman/troubleshooting.html#recrawl
```

Reviewers: fsilva, sid0

Reviewed By: sid0

Subscribers: sid0, jimp, dreiss

Differential Revision: https://phabricator.fb.com/D1488381