Add get-config command
Summary:
the Buck folks would like to be able to detect whether the
buck-out dir is correctly listed in the ignore_dirs list in the
.watchmanconfig file.

This diff adds a mechanism for querying watchman to get the
configuration that is in use for a watch.

Test Plan:
`make integration`.  Also tested manually; the output from those
tests is reproduced in the documentation.

Reviewers: fsilva, sid0

Reviewed By: sid0

Subscribers: jimp, dreiss

Differential Revision: https://phabricator.fb.com/D1488644